Don't package base library
The solution is split into a core project and an ASP.NET Core project.
This is looking forward to when EF Core support comes in and we will
have a common 'core' library with on-demand integration with ASP.NET
Core and EF Core.

The down-side of this split is the need to publish multiple packages
and the associated versioning challenges (see
https://stackoverflow.com/questions/53278544/nuget-versioning-with-projectreference-dependencies).

For now include the core type as source into the AspNetCore project so
we can publish a single package.